Update main
cd to $PBS_O_WORKDIR is not necessar when it's run through the default hook as it sets `-d $PWD` when it tries to qsub. Also, it will cause workdir to reset to $HOME if $PBS_O_WORKDIR is not set (like on slurm). So I've removed it.

Instead of looking for t1.nii.gz, the return code from N4BiasFieldCorrection should be checked to see if the application finished successfully. Since bash passes the return code of the last command to parent script, so as long as N4BiasFieldCorrection is the last command on this script, you don't have to check it manually. Also, you don't have to write "finished" as it is already taken care of by the default hook.